backend/requirements.txt
anyio==3.6.2
asgiref==3.7.2
# backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==3.2.25
django-cors-headers==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
dnspython==2.3.0
drf-yasg==1.21.7
email-validator==2.0.0.post2
fastapi==0.99.0
flake8==5.0.4
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.7
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
orjson==3.9.1
packaging==24.0
pycparser==2.21
pydantic==1.10.10
PyJWT==2.8.0
pyTelegramBotAPI==4.10.0
python-dotenv==0.19.0
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
SQLAlchemy==2.0.8
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.2.3
WTForms==3.0.1
yarg==0.1.9
zipp==3.15.0
